backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.1
sqlalchemy==2.0.40
pydantic==2.11.3
pydantic-settings==2.8.1
psycopg2-binary
python-dotenv
python-multipart==0.0.20
pandas
openpyxl==3.1.5
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
python-magic==0.4.27
bcrypt==4.2.1
httpx==0.28.1
